Extended functionality and polished ouput (#5)
* Extended functionality and polished ouput

- added the notification options to the template
- removed notification options from the file screen
- added active/passiv screen type configuration
- added german translation to the notification messages
- output polished a bit ;-)
- added a few more comments

* changed default language to english

* changed default type to active

+++ DeployCryptoBlocker.ps1
@@ -1,5 +1,40 @@
 # DeployCryptoBlocker.ps1
-#
+# Version: 1.1
+#####
+
+################################ USER CONFIGURATION ################################
+
+# Names to use in FSRM
+$fileGroupName = "CryptoBlockerGroup"
+$fileTemplateName = "CryptoBlockerTemplate"
+# set screening type to
+# Active screening: Do not allow users to save unathorized files
+$fileTemplateType = "Active"
+# Passive screening: Allow users to save unathorized files (use for monitoring)
+#$fileTemplateType = "Passiv"
+
+# Write the email options to the temporary file - comment out the entire block if no email notification should be set
+$EmailNotification = $env:TEMP + "\tmpEmail001.tmp"
+"Notification=m" >> $EmailNotification
+"To=[Admin Email]" >> $EmailNotification
+## en
+"Subject=Unauthorized file from the [Violated File Group] file group detected" >> $EmailNotification
+"Message=User [Source Io Owner] attempted to save [Source File Path] to [File Screen Path] on the [Server] server. This file is in the [Violated File Group] file group, which is not permitted on the server."  >> $EmailNotification
+## de
+#"Subject=Nicht autorisierte Datei erkannt, die mit Dateigruppe [Violated File Group] übereinstimmt" >> $EmailNotification
+#"Message=Das System hat erkannt, dass Benutzer [Source Io Owner] versucht hat, die Datei [Source File Path] unter [File Screen Path] auf Server [Server] zu speichern. Diese Datei weist Übereinstimmungen mit der Dateigruppe [Violated File Group] auf, die auf dem System nicht zulässig ist."  >> $EmailNotification
+
+# Write the event log options to the temporary file - comment out the entire block if no event notification should be set
+$EventNotification = $env:TEMP + "\tmpEvent001.tmp"
+"Notification=e" >> $EventNotification
+"EventType=Warning" >> $EventNotification
+## en
+"Message=User [Source Io Owner] attempted to save [Source File Path] to [File Screen Path] on the [Server] server. This file is in the [Violated File Group] file group, which is not permitted on the server." >> $EventNotification
+## de
+#"Message=Das System hat erkannt, dass Benutzer [Source Io Owner] versucht hat, die Datei [Source File Path] unter [File Screen Path] auf Server [Server] zu speichern. Diese Datei weist Übereinstimmungen mit der Dateigruppe [Violated File Group] auf, die auf dem System nicht zulässig ist." >> $EventNotification
+
+################################ USER CONFIGURATION ################################
+
 ################################ Functions ################################
 
 Function ConvertFrom-Json20
@@ -73,6 +108,8 @@ Function New-CBArraySplit
 
 ################################ Functions ################################
 
+################################ Program code ################################
+
 # Get all drives with shared folders, these drives will get FRSRM protection
 $DrivesContainingShares = @(Get-WmiObject Win32_Share |            # all shares on this computer, filter:
                             Where-Object { $_.Type -eq 0 } |       # 0 = disk drives (not printers, IPC$, C$ Admin shares)
@@ -84,17 +121,20 @@ $DrivesContainingShares = @(Get-WmiObject Win32_Share |            # all shares 
 
 if ($drivesContainingShares.Count -eq 0)
 {
+    Write-Host "`n####"
     Write-Host "No drives containing shares were found. Exiting.."
     exit
 }
 
+Write-Host "`n####"
 Write-Host "The following shares needing to be protected: $($drivesContainingShares -Join ",")"
 
 
-#### Identify Windows Server version, and install FSRM role
+# Identify Windows Server version, and install FSRM role
 $majorVer = [System.Environment]::OSVersion.Version.Major
 $minorVer = [System.Environment]::OSVersion.Version.Minor
 
+Write-Host "`n####"
 Write-Host "Checking File Server Resource Manager.."
 
 Import-Module ServerManager
@@ -106,39 +146,43 @@ if ($majorVer -ge 6)
     if ($minorVer -ge 2 -and $checkFSRM.Installed -ne "True")
     {
         # Server 2012
+        Write-Host "`n####"
         Write-Host "FSRM not found.. Installing (2012).."
         Install-WindowsFeature -Name FS-Resource-Manager -IncludeManagementTools
     }
     elseif ($minorVer -ge 1 -and $checkFSRM.Installed -ne "True")
     {
         # Server 2008 R2
-        Write-Host "FSRM not found.. Installing (2008 R2).."
+        Write-Host "`n####"
+		Write-Host "FSRM not found.. Installing (2008 R2).."
         Add-WindowsFeature FS-FileServer, FS-Resource-Manager
     }
     elseif ($checkFSRM.Installed -ne "True")
     {
         # Server 2008
-        Write-Host "FSRM not found.. Installing (2008).."
+        Write-Host "`n####"
+		Write-Host "FSRM not found.. Installing (2008).."
         &servermanagercmd -Install FS-FileServer FS-Resource-Manager
     }
 }
 else
 {
     # Assume Server 2003
-    Write-Host "Unsupported version of Windows detected! Quitting.."
+    Write-Host "`n####"
+	Write-Host "Unsupported version of Windows detected! Quitting.."
     return
 }
 
-
-$fileGroupName = "CryptoBlockerGroup"
-$fileTemplateName = "CryptoBlockerTemplate"
-$fileScreenName = "CryptoBlockerScreen"
-
 # Download list of CryptoLocker file extensions
+Write-Host "`n####"
+Write-Host "Dowloading CryptoLocker file extensions list from fsrm.experiant.ca api.."
 $webClient = New-Object System.Net.WebClient
 $jsonStr = $webClient.DownloadString("https://fsrm.experiant.ca/api/v1/get")
 $monitoredExtensions = @(ConvertFrom-Json20 $jsonStr | ForEach-Object { $_.filters })
 
+# Process SkipList.txt
+Write-Host "`n####"
+Write-Host "Processing SkipList.."
 If (Test-Path .\SkipList.txt)
 {
     $Exclusions = Get-Content .\SkipList.txt | ForEach-Object { $_.Trim() }
@@ -164,47 +208,57 @@ Else
     Set-Content -Path .\SkipList.txt -Value $emptyFile
 }
 
-
 # Split the $monitoredExtensions array into fileGroups of less than 4kb to allow processing by filescrn.exe
 $fileGroups = @(New-CBArraySplit $monitoredExtensions)
 
 # Perform these steps for each of the 4KB limit split fileGroups
+Write-Host "`n####"
+Write-Host "Adding/replacing File Groups.."
 ForEach ($group in $fileGroups) {
-    Write-Host "Adding/replacing File Group [$($group.fileGroupName)] with monitored file [$($group.array -Join ",")].."
-    &filescrn.exe filegroup Delete "/Filegroup:$($group.fileGroupName)" /Quiet
+    #Write-Host "Adding/replacing File Group [$($group.fileGroupName)] with monitored file [$($group.array -Join ",")].."
+    Write-Host "`nFile Group [$($group.fileGroupName)] with monitored files from [$($group.array[0])] to [$($group.array[$group.array.GetUpperBound(0)])].."
+	&filescrn.exe filegroup Delete "/Filegroup:$($group.fileGroupName)" /Quiet
     &filescrn.exe Filegroup Add "/Filegroup:$($group.fileGroupName)" "/Members:$($group.array -Join '|')"
 }
 
-Write-Host "Adding/replacing File Screen Template [$fileTemplateName] with Event Notification [$eventConfFilename] and Command Notification [$cmdConfFilename].."
+# Create File Screen Template with Notification
+Write-Host "`n####"
+Write-Host "Adding/replacing [$fileTemplateType] File Screen Template [$fileTemplateName] with eMail Notification [$EmailNotification] and Event Notification [$EventNotification].."
 &filescrn.exe Template Delete /Template:$fileTemplateName /Quiet
-# Build the argument list with all required fileGroups
-$screenArgs = 'Template', 'Add', "/Template:$fileTemplateName"
+# Build the argument list with all required fileGroups and notifications
+$screenArgs = 'Template', 'Add', "/Template:$fileTemplateName", "/Type:$fileTemplateType"
 ForEach ($group in $fileGroups) {
     $screenArgs += "/Add-Filegroup:$($group.fileGroupName)"
 }
-
+If ($EmailNotification -ne "") {
+    $screenArgs += "/Add-Notification:m,$EmailNotification"
+}
+If ($EventNotification -ne "") {
+    $screenArgs += "/Add-Notification:e,$EventNotification"
+}
 &filescrn.exe $screenArgs
 
-$EmailNotification = $env:TEMP + "\tmpEmail001.tmp"
-$EventNotification = $env:TEMP + "\tmpEvent001.tmp"
-
-# Write the email options to the temporary file
-"Notification=m" >> $EmailNotification
-"To=[Admin Email]" >> $EmailNotification
-"Subject=Unauthorized file from the [Violated File Group] file group detected" >> $EmailNotification
-"Message=User [Source Io Owner] attempted to save [Source File Path] to [File Screen Path] on the [Server] server. This file is in the [Violated File Group] file group, which is not permitted on the server."  >> $EmailNotification
-
-# Write the event log options to the temporary file
-"Notification=e" >> $EventNotification
-"EventType=Warning" >> $EventNotification
-"Message=User [Source Io Owner] attempted to save [Source File Path] to [File Screen Path] on the [Server] server. This file is in the [Violated File Group] file group, which is not permitted on the server." >> $EventNotification
-
+# Create File Screens for every drive containing shares
+Write-Host "`n####"
 Write-Host "Adding/replacing File Screens.."
 $drivesContainingShares | ForEach-Object {
-    Write-Host "`tAdding/replacing File Screen for [$_] with Source Template [$fileTemplateName].."
+    Write-Host "File Screen for [$_] with Source Template [$fileTemplateName].."
     &filescrn.exe Screen Delete "/Path:$_" /Quiet
-    &filescrn.exe Screen Add "/Path:$_" "/SourceTemplate:$fileTemplateName" "/Add-Notification:m,$EmailNotification" "/Add-Notification:e,$EventNotification"
+    &filescrn.exe Screen Add "/Path:$_" "/SourceTemplate:$fileTemplateName"
 }
 
-Remove-Item $EmailNotification -Force
-Remove-Item $EventNotification -Force
+# Cleanup temporary files if they were created
+Write-Host "`n####"
+Write-Host "Cleaning up temporary stuff.."
+If ($EmailNotification -ne "") {
+	Remove-Item $EmailNotification -Force
+}
+If ($EventNotification -ne "") {
+	Remove-Item $EventNotification -Force
+}
+
+Write-Host "`n####"
+Write-Host "Done."
+Write-Host "####"
+
+################################ Program code ################################
